Uploaded old Catch2 test files...
IMPORTANT:  These files include:
a) The catch.hpp file, which contains all the test commands (all implemented as template functions.  This is an older version and should probably be updated, however,
b) Catch2 test files that were used on an earlier version of a Date class.  They are included as a guide, but to reuse them, the namespace for the date functionality in the target code will need to be updated to "Dates", and the "Date" class to "XLDate".
In their present form, they will not run with the current XLDate and DayCount code.